pi_client/requirements.txt
httpx>=0.27.0
pillow>=10.0.0
inky>=2.0.0
uvicorn>=0.32.0
starlette>=0.41.0
